lib: in sci changed upon to comments,  moved config file to the sci folder
