+++ src/lib/sci/config.ts
